+++ src/mixin/class.rs
@@ -422,7 +422,7 @@ impl Class {
     /// ```
     ///
     // Link to docs.rs since `Class` may either be in `class` module or root
-    /// [`def_method!`]: https://docs.rs/rosy/0.0.4/rosy/macro.def_method.html
+    /// [`def_method!`]: https://docs.rs/rosy/0.0.5/rosy/macro.def_method.html
     pub fn def_method<N, F>(self, name: N, f: F) -> Result
     where
         N: Into<SymbolId>,
@@ -444,7 +444,7 @@ impl Class {
     /// exception will be raised.
     ///
     // Link to docs.rs since `Class` may either be in `class` module or root
-    /// [`def_method_unchecked!`]: https://docs.rs/rosy/0.0.4/rosy/macro.def_method_unchecked.html
+    /// [`def_method_unchecked!`]: https://docs.rs/rosy/0.0.5/rosy/macro.def_method_unchecked.html
     #[inline]
     pub unsafe fn def_method_unchecked<N, F>(self, name: N, f: F)
     where