+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 [Semantic Versioning].
 
 ## [Unreleased]
+
+## [0.0.5] - 2019-05-20
 ### Added
 - Variants of `Class::new_instance` that are `unsafe` or take arguments
 - `PartialEq<[A]>` implementation for `Array<O>` where `O: PartialEq<A>`
@@ -69,7 +71,8 @@ Initial release
 [Keep a Changelog]:    http://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html
 
-[Unreleased]: https://github.com/oceanpkg/rosy/compare/v0.0.4...HEAD
+[Unreleased]: https://github.com/oceanpkg/rosy/compare/v0.0.5...HEAD
+[0.0.5]: https://github.com/oceanpkg/rosy/compare/v0.0.4...v0.0.5
 [0.0.4]: https://github.com/oceanpkg/rosy/compare/v0.0.3...v0.0.4
 [0.0.3]: https://github.com/oceanpkg/rosy/compare/v0.0.2...v0.0.3
 [0.0.2]: https://github.com/oceanpkg/rosy/compare/v0.0.1...v0.0.2